Allows decimals to be entered

+++ app/forms/feedback_forms.py
@@ -2,6 +2,7 @@ from flask_babel import gettext
 from flask_wtf import FlaskForm
 from wtforms import HiddenField
 from wtforms import IntegerField
+from wtforms import FloatField
 from wtforms import RadioField
 from wtforms import TextAreaField
 from wtforms.validators import DataRequired
@@ -146,7 +147,7 @@ class EndOfApplicationPage3Form(EndOfApplicationPageForm):
 
 
 class EndOfApplicationPage4Form(EndOfApplicationPageForm):
-    hours_spent = IntegerField(
+    hours_spent = FloatField(
         label="Number of hours spent:",
         validators=[
             DataRequired(message="Enter a number only. The number must be at least 0.5 or greater."),
